splitted env vars into env.client and env.server and updated the imports
