+++ frontend/src/app/auth/login/page.tsx
@@ -1,4 +1,4 @@
-import { isGithubAuthEnabled } from 'utils/credentials'
+import { isGithubAuthEnabled } from 'utils/env.server'
 import LoginPageContent from 'components/LoginPageContent'
 
 export default function LoginPage() {
