+++ frontend/src/app/layout.tsx
@@ -3,7 +3,8 @@ import type { Metadata } from 'next'
 import { Geist, Geist_Mono } from 'next/font/google'
 import React from 'react'
 import { Providers } from 'wrappers/provider'
-import { GTM_ID, isGithubAuthEnabled } from 'utils/credentials'
+import { GTM_ID } from 'utils/env.client'
+import { isGithubAuthEnabled } from 'utils/env.server'
 import AutoScrollToTop from 'components/AutoScrollToTop'
 import BreadCrumbs from 'components/BreadCrumbs'
 import Footer from 'components/Footer'
