+++ frontend/src/components/CardDetailsPage.tsx
@@ -14,7 +14,7 @@ import { useRouter } from 'next/navigation'
 import { useSession } from 'next-auth/react'
 import type { ExtendedSession } from 'types/auth'
 import type { DetailsCardProps } from 'types/card'
-import { IS_PROJECT_HEALTH_ENABLED } from 'utils/credentials'
+import { IS_PROJECT_HEALTH_ENABLED } from 'utils/env.client'
 import { scrollToAnchor } from 'utils/scrollToAnchor'
 import { getSocialIcon } from 'utils/urlIconMappings'
 import AnchorTitle from 'components/AnchorTitle'
@@ -87,7 +87,7 @@ const DetailsCard = ({
               ) && (
                 <button
                   type="button"
-                  className="flex items-center justify-center gap-2 text-nowrap rounded-md border border-[#0D6EFD] bg-transparent px-2 py-2 text-[#0D6EFD] text-blue-600 transition-all hover:bg-[#0D6EFD] hover:text-white dark:border-sky-600 dark:text-sky-600 dark:hover:bg-sky-100"
+                  className="flex items-center justify-center gap-2 text-nowrap rounded-md border border-[#0D6EFD] bg-transparent px-2 py-2 text-[#0D6EFD] transition-all hover:bg-[#0D6EFD] hover:text-white dark:border-sky-600 dark:text-sky-600 dark:hover:bg-sky-100"
                   onClick={() => {
                     router.push(`${window.location.pathname}/edit`)
                   }}
@@ -163,24 +163,24 @@ const DetailsCard = ({
             type === 'committee' ||
             type === 'user' ||
             type === 'organization') && (
-            <SecondaryCard
-              icon={faChartPie}
-              title={<AnchorTitle title="Statistics" />}
-              className="md:col-span-2"
-            >
-              {stats.map((stat, index) => (
-                <div key={index}>
-                  <InfoBlock
-                    className="pb-1"
-                    icon={stat.icon}
-                    pluralizedName={stat.pluralizedName}
-                    unit={stat.unit}
-                    value={stat.value}
-                  />
-                </div>
-              ))}
-            </SecondaryCard>
-          )}
+              <SecondaryCard
+                icon={faChartPie}
+                title={<AnchorTitle title="Statistics" />}
+                className="md:col-span-2"
+              >
+                {stats.map((stat, index) => (
+                  <div key={index}>
+                    <InfoBlock
+                      className="pb-1"
+                      icon={stat.icon}
+                      pluralizedName={stat.pluralizedName}
+                      unit={stat.unit}
+                      value={stat.value}
+                    />
+                  </div>
+                ))}
+              </SecondaryCard>
+            )}
           {type === 'chapter' && geolocationData && (
             <div className="mb-8 h-[250px] md:col-span-4 md:h-auto">
               <ChapterMapWrapper
@@ -262,31 +262,31 @@ const DetailsCard = ({
           type === 'repository' ||
           type === 'user' ||
           type === 'organization') && (
-          <div className="grid-cols-2 gap-4 lg:grid">
-            <RecentIssues data={recentIssues} showAvatar={showAvatar} />
-            {type === 'user' ||
-            type === 'organization' ||
-            type === 'repository' ||
-            type === 'project' ? (
-              <Milestones data={recentMilestones} showAvatar={showAvatar} />
-            ) : (
-              <RecentReleases
-                data={recentReleases}
-                showAvatar={showAvatar}
-                showSingleColumn={true}
-              />
-            )}
-          </div>
-        )}
+            <div className="grid-cols-2 gap-4 lg:grid">
+              <RecentIssues data={recentIssues} showAvatar={showAvatar} />
+              {type === 'user' ||
+                type === 'organization' ||
+                type === 'repository' ||
+                type === 'project' ? (
+                <Milestones data={recentMilestones} showAvatar={showAvatar} />
+              ) : (
+                <RecentReleases
+                  data={recentReleases}
+                  showAvatar={showAvatar}
+                  showSingleColumn={true}
+                />
+              )}
+            </div>
+          )}
         {(type === 'project' ||
           type === 'repository' ||
           type === 'organization' ||
           type === 'user') && (
-          <div className="grid-cols-2 gap-4 lg:grid">
-            <RecentPullRequests data={pullRequests} showAvatar={showAvatar} />
-            <RecentReleases data={recentReleases} showAvatar={showAvatar} showSingleColumn={true} />
-          </div>
-        )}
+            <div className="grid-cols-2 gap-4 lg:grid">
+              <RecentPullRequests data={pullRequests} showAvatar={showAvatar} />
+              <RecentReleases data={recentReleases} showAvatar={showAvatar} showSingleColumn={true} />
+            </div>
+          )}
         {(type === 'project' || type === 'user' || type === 'organization') &&
           repositories.length > 0 && (
             <SecondaryCard icon={faFolderOpen} title={<AnchorTitle title="Repositories" />}>
