+++ frontend/src/utils/env.client.ts
@@ -1,0 +1,10 @@
+export const API_URL = process.env.NEXT_PUBLIC_API_URL
+export const CSRF_URL = process.env.NEXT_PUBLIC_CSRF_URL
+export const ENVIRONMENT = process.env.NEXT_PUBLIC_ENVIRONMENT
+export const GRAPHQL_URL = process.env.NEXT_PUBLIC_GRAPHQL_URL
+export const GTM_ID = process.env.NEXT_PUBLIC_GTM_ID
+export const IDX_URL = process.env.NEXT_PUBLIC_IDX_URL
+export const IS_PROJECT_HEALTH_ENABLED =
+    process.env.NEXT_PUBLIC_IS_PROJECT_HEALTH_ENABLED === 'true'
+export const RELEASE_VERSION = process.env.NEXT_PUBLIC_RELEASE_VERSION
+export const SENTRY_DSN = process.env.NEXT_PUBLIC_SENTRY_DSN
