+++ frontend/src/utils/helpers/apolloClient.ts
@@ -1,7 +1,7 @@
 import { ApolloClient, createHttpLink, InMemoryCache } from '@apollo/client'
 import { setContext } from '@apollo/client/link/context'
 import { AppError, handleAppError } from 'app/global-error'
-import { GRAPHQL_URL } from 'utils/credentials'
+import { GRAPHQL_URL } from 'utils/env.client'
 import { getCsrfToken } from 'utils/utility'
 const createApolloClient = () => {
   if (!GRAPHQL_URL) {
